riorganizzazione locazione funzioni di analisi rumore
